+++ api/src/main/kotlin/com/inner/circle/api/controller/merchant/MerchantPaymentController.kt
@@ -7,10 +7,13 @@ import com.inner.circle.api.config.SwaggerConfig
 import com.inner.circle.api.controller.PaymentForMerchantV1Api
 import com.inner.circle.api.controller.dto.PaymentApproveDto
 import com.inner.circle.api.controller.dto.PaymentResponse
+import com.inner.circle.api.controller.dto.TransactionDto
 import com.inner.circle.api.controller.dto.UserCardDto
+import com.inner.circle.api.controller.request.CancelPaymentRequest
 import com.inner.circle.api.controller.request.PaymentApproveRequest
 import com.inner.circle.api.controller.request.PaymentClaimRequest
 import com.inner.circle.core.security.MerchantUserDetails
+import com.inner.circle.core.usecase.CancelPaymentUseCase
 import com.inner.circle.core.usecase.PaymentClaimUseCase
 import com.inner.circle.core.usecase.SavePaymentApproveUseCase
 import com.inner.circle.core.usecase.UserCardUseCase
@@ -22,6 +25,7 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 
@@ -32,7 +36,8 @@ class MerchantPaymentController(
     private val claimUseCase: PaymentClaimUseCase,
     private val savePaymentApproveService: SavePaymentApproveUseCase,
     private val userCardUseCase: UserCardUseCase,
-    private val statusChangedMessageSender: PaymentStatusChangedMessageSender
+    private val statusChangedMessageSender: PaymentStatusChangedMessageSender,
+    private val cancelPaymentUseCase: CancelPaymentUseCase
 ) {
     private val logger: Logger = LoggerFactory.getLogger(MerchantPaymentController::class.java)
 
@@ -96,6 +101,25 @@ class MerchantPaymentController(
         return PaymentResponse.ok(
             data
         )
+    }
+
+    @Operation(summary = "결제 취소")
+    @PostMapping("/payments/{paymentKey}/cancel")
+    fun cancelPayment(
+        @AuthenticationPrincipal merchantUserDetails: MerchantUserDetails,
+        @PathVariable("paymentKey") paymentKey: String,
+        @Valid @RequestBody request: CancelPaymentRequest
+    ): PaymentResponse<TransactionDto> {
+        val transaction =
+            cancelPaymentUseCase.cancel(
+                CancelPaymentUseCase.CancelPaymentRequest(
+                    merchantId = merchantUserDetails.getId(),
+                    paymentKey = paymentKey,
+                    amount = request.amount
+                )
+            )
+
+        return PaymentResponse.ok(TransactionDto.of(transaction))
     }
 
     private fun sendStatusChangedMessage(
