+++ api/src/test/resources/sql/schema.sql
@@ -99,8 +99,8 @@ create table if not exists account
 );
 
 INSERT INTO account (id, email, password, status, created_at, updated_at) VALUES (293847562342874239, 'test@test.com', '$2a$10$8I6102l2DJvon6gkFJUHh.ZAeszd5swcabHdpr8iFsDXJ6WOZ51v2', 'ACTIVE', '2025-02-12 22:05:32.901464', '2025-02-12 22:05:32.901464');
-INSERT INTO merchant (id, email, password, name, created_at, updated_at) VALUES (679851028133504349, 'pay200', 'pay200', 'pay200', '2025-02-11 20:25:42.000000', '2025-02-11 20:25:42.000000');
-INSERT INTO api_key (id, merchant_id, token, created_at, updated_at) VALUES (679851702594364766, 679851028133504349, 'pay200', '2025-02-11 20:25:42.000000', '2025-02-11 20:25:42.000000');
+INSERT INTO merchant (id, email, password, name, created_at, updated_at) VALUES (680785231122250286, 'pay200-shop@pay.com', '$2a$10$hk5caerjOASeaFaKPCwcZebXd/YRTcMJoAs0/sikxkvB6BQgZ5w7G', '테스트상점', '2025-02-11 20:25:42.000000', '2025-02-11 20:25:42.000000');
+INSERT INTO api_key (id, merchant_id, token, created_at, updated_at) VALUES (679851702594364766, 680785231122250286, 'ilfwYbPuUuXl-cyWCB7J390A5IfLi7fSu7ZNXD4kreM=', '2025-02-11 20:25:42.000000', '2025-02-11 20:25:42.000000');
 INSERT INTO user_card (id, account_id, is_representative, card_number, expiration_period, cvc, card_company, created_at, updated_at) VALUES (7295051915259393268, 5, true, '4567-8923-6378-3982', '03/29', '654', '삼성', '2025-02-11 21:20:14.188948', '2025-02-11 21:20:14.188948');
 INSERT INTO user_card (id, account_id, is_representative, card_number, expiration_period, cvc, card_company, created_at, updated_at) VALUES (7295051915258292529, 3, true, '9876-5432-1098-7654', '01/27', '789', '신한', '2025-02-11 21:20:14.188948', '2025-02-11 21:20:14.188948');
 INSERT INTO user_card (id, account_id, is_representative, card_number, expiration_period, cvc, card_company, created_at, updated_at) VALUES (7295051915259438759, 2, true, '5678-1234-5678-9012', '10/26', '456', '현대', '2025-02-11 21:20:14.188948', '2025-02-11 21:20:14.188948');
