+++ infra/src/main/kotlin/com/inner/circle/infra/repository/PaymentJpaRepository.kt
@@ -4,8 +4,8 @@ import com.inner.circle.infra.repository.entity.PaymentEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PaymentJpaRepository : JpaRepository<PaymentEntity, String> {
-    fun findByAccountIdAndPaymentKey(
-        accountId: Long,
+    fun findByMerchantIdAndPaymentKey(
+        merchantId: Long,
         paymentKey: String
     ): PaymentEntity?
 }
